DQ: Started implementing labels
I've copied [tsoding](https://github.com/tsoding/deque.py)'s idea of
using labels, but it is still a mess to do.
Maybe I should think more on how to call on the address of the label,
and not to parse instructions that is inside of a label.
Whish me luck :^)